Rmb/routing updates (#175)
1. add a response for this
`/.well-known/appspecific/com.chrome.devtools.json` path that chrome dev
tools seems to be requesting. not sure what this is all about.
2. fix layout related bugs: 1) navigating directly to a path like
`/profile` was raising an error due to viewer not being present. This is
resolved with the `result.loading || !viewer` check. 2) the `useEffect`
blocks were executing twice, resulting in the `returnTo` not being set
properly.
3. implements the `returnTo` functionality via `localStorage`. If a user
visits a path and gets redirected to login, the path will be remembered
and redirected to after sign in
4. dont redirect admins to the admin panel on login - its confusing.
instead expose an admin link in the sidebar